+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Added `Insert chapter marker` button in text editor to insert chapter markers at the current cursor position.
+- Added `Preview` button in voice mixer to preview the voice mix with the selected settings.
 - Fixed `f-string: unmatched '['` error in Voice preview, mentioned in #14
-- Fixed the content before first chapter not being included in the output (added handling for initial content).
+- Fixed the issue with the content before first chapter not being included in the output.
 - Fixed m4b chapter generation opens CMD window in Windows.